requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
flake8==3.5.0
flake8-quotes==0.14.0
Flask==0.12.2
googlemaps==2.5.1
haversine==0.4.5
idna==2.6
mccabe==0.6.1
nose==1.3.7
numpy==1.14.2
pandas==0.22.0
parameterized==0.6.1
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.4.1
python-dateutil==2.6.1
requests==2.18.4
requests-cache==0.4.13
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tqdm==4.19.5
untangle==1.1.1
urllib3==1.22
vulture==0.26
weightreservoir==1.0
